Refactor normalizeWidgetData using single responsibility principle
